SSCSSI-442: Remove feature toggles related to concurrency change - HEARINGS_CASE_UPDATE_V2_ENABLED
