+++ charts/sscs-tribunals-api/Chart.yaml
@@ -1,7 +1,7 @@
 apiVersion: v2
 name: sscs-tribunals-api
 home: https://github.com/hmcts/sscs-tribunals-case-api
-version: 0.0.172
+version: 0.0.173
 description: SSCS Tribunals Case API
 maintainers:
   - name: HMCTS SSCS Team
