+++ charts/sscs-tribunals-api/values.hearings.preview.template.yaml
@@ -82,7 +82,6 @@ java:
     HMC_DEPLOYMENT_ID: "deployment-${SERVICE_NAME}"
     HMC_DEPLOYMENT_FILTER_ENABLED: true
     ADJOURNMENT_ENABLED: true
-    BYPASS_HEARING_API_SERVICE: true
   secrets:
     JOB_SCHEDULER_DB_PASSWORD:
       secretRef: postgres
