+++ charts/sscs-tribunals-api/values.preview.template.yaml
@@ -38,7 +38,6 @@ java:
     ELINKS_V2_FEATURE_ENABLED: true
     DELETED_REDACTED_DOC_ENABLED: true
     BYPASS_NOTIFICATIONS_SERVICE: true
-    BYPASS_HEARING_API_SERVICE: true
     TESTING_SUPPORT_ENABLED: true
     EXUI_API_URL: "https://${SERVICE_NAME}-xui"
     TOPIC_NAME: "${SERVICE_NAME}-servicebus-main-topic"
@@ -63,7 +62,6 @@ java:
     RUN_DB_MIGRATION_ON_STARTUP: true
     TEST_RECIPIENTS_POSTCODE: "*"
     INFECTED_BLOOD_COMPENSATION_ENABLED: true
-    PROCESS_EVENT_MESSAGE_V2_ENABLED: true
     # START - SSCS-IB RELEASE VARIABLES
     DOCUMENTS_ENGLISH_DIRECTION_ISSUED: TB-SCS-GNO-ENG-directions-notice-v2.docx
     DOCUMENTS_WELSH_DIRECTION_ISSUED: TB-SCS-GNO-ENG-directions-notice-v2.docx
