+++ definitions/benefit/sheets/CaseEvent/CaseEvent.json
@@ -3269,5 +3269,15 @@
     "PreConditionState(s)": "*",
     "PostConditionState": "*",
     "SecurityClassification": "Public"
+  },
+  {
+    "LiveFrom": "17/01/2025",
+    "CaseTypeID": "Benefit",
+    "ID": "clearExpiredFilters",
+    "Name": "Cleared expired filters",
+    "DisplayOrder": 1302,
+    "PreConditionState(s)": "*",
+    "PostConditionState": "*",
+    "SecurityClassification": "Public"
   }
 ]
