+++ definitions/benefit/sheets/ComplexTypes/ComplexTypes.json
@@ -312,7 +312,7 @@
   {"LiveFrom": "30/07/2018", "ID": "ScannedDocument", "ListElementCode": "includeInBundle", "FieldType": "YesOrNo", "ElementLabel": "Include in bundle?", "SecurityClassification": "PUBLIC"},
   {"LiveFrom": "01/01/2018", "ID": "ScannedDocument", "ListElementCode": "originalSenderOtherPartyId", "FieldType": "Text", "ElementLabel": "Original sender other party ID", "FieldShowCondition": "type=\"AnyValueToFailThisCondition\"", "SecurityClassification": "PUBLIC"},
   {"LiveFrom": "01/01/2018", "ID": "ScannedDocument", "ListElementCode": "originalSenderOtherPartyName", "FieldType": "Text", "ElementLabel": "Original sender other party name", "SecurityClassification": "PUBLIC"},
-  {"LiveFrom": "18/06/2019", "ID": "dwpResponseDocumentCT", "ListElementCode": "dwpResponseDocumentTitle", "FieldType": "Label", "ElementLabel": "Secretary of State's response", "SecurityClassification": "PUBLIC"},
+  {"LiveFrom": "18/06/2019", "ID": "dwpResponseDocumentCT", "ListElementCode": "dwpResponseDocumentTitle", "FieldType": "Label", "ElementLabel": "Secretary of State's response", "SecurityClassification": "PUBLIC", "FieldShowCondition": "documentFileName=\"AnyValueToFailThisCondition\""},
   {"LiveFrom": "18/06/2019", "ID": "dwpResponseDocumentCT", "ListElementCode": "hmrcResponseDocumentTitle", "FieldType": "Label", "ElementLabel": "HM Revenue and Customs", "SecurityClassification": "PUBLIC", "FieldShowCondition": "documentFileName=\"AnyValueToFailThisCondition\""},
   {"LiveFrom": "18/06/2019", "ID": "dwpResponseDocumentCT", "ListElementCode": "ibcaResponseDocumentTitle", "FieldType": "Label", "ElementLabel": "Infected Blood Compensation Authority's response", "SecurityClassification": "PUBLIC", "FieldShowCondition": "documentFileName=\"AnyValueToFailThisCondition\""},
   {"LiveFrom": "18/06/2019", "ID": "dwpResponseDocumentCT", "ListElementCode": "documentFileName", "FieldType": "Text", "ElementLabel": "Evidence bundle file name", "FieldShowCondition": "dwpResponseDocumentTitle=\"AnyValueToFailThisCondition\"", "SecurityClassification": "PUBLIC"},
