+++ src/IntegrationTests/resources/config/application_it.properties
@@ -55,5 +55,3 @@ hmc.url=http://localhost:10010
 
 flags.postHearings.enabled=false
 feature.elinksV2.enabled=false
-feature.bypass-hearing-api-service.enabled=true
-
