+++ src/IntegrationTests/resources/config/application_it_updateCaseV2.properties
@@ -4,5 +4,3 @@ flags.postHearings.enabled=false
 feature.elinksV2.enabled=false
 
 flags.tribunals-to-hearings-api.enabled=true
-feature.bypass-hearing-api-service.enabled=true
-feature.process-event-message-v2.enabled=true
