+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/readytolist/ReadyToListAboutToSubmitHandler.java
@@ -20,6 +20,10 @@ import uk.gov.hmcts.reform.sscs.service.servicebus.HearingMessagingServiceFactor
 @Slf4j
 public class ReadyToListAboutToSubmitHandler implements PreSubmitCallbackHandler<SscsCaseData> {
 
+    static final String EXISTING_HEARING_WARNING = "There is already a hearing request in List assist, "
+            + "are you sure you want to send another request? If you do proceed, then please cancel the existing hearing request first";
+    static final String GAPS_CASE_WARNING = "This is a GAPS case, If you do want to proceed, "
+            + "then please change the hearing route to List Assist";
     private final boolean gapsSwitchOverFeature;
 
     private final RegionalProcessingCenterService regionalProcessingCenterService;
@@ -56,10 +60,8 @@ public class ReadyToListAboutToSubmitHandler implements PreSubmitCallbackHandler
 
             if (!callback.isIgnoreWarnings() && !YesNo.YES.equals(sscsCaseData.getIgnoreCallbackWarnings())) {
                 PreSubmitCallbackResponse<SscsCaseData> response = new PreSubmitCallbackResponse<>(callback.getCaseDetails().getCaseData());
-                String gapsProceedWarning = "This is a GAPS case, If you do want to proceed, "
-                    + "then please change the hearing route to List Assist";
-                response.addWarning(gapsProceedWarning);
-                log.warn("Warning: {}", gapsProceedWarning);
+                response.addWarning(GAPS_CASE_WARNING);
+                log.warn("Warning: {}", GAPS_CASE_WARNING);
                 return response;
             }
 
@@ -68,12 +70,10 @@ public class ReadyToListAboutToSubmitHandler implements PreSubmitCallbackHandler
         }
 
         if (SscsHelper.hasHearingScheduledInTheFuture(sscsCaseData)
-                && !callback.isIgnoreWarnings()) {
+                && !callback.isIgnoreWarnings() && !YesNo.YES.equals(sscsCaseData.getIgnoreCallbackWarnings())) {
             PreSubmitCallbackResponse<SscsCaseData> response = new PreSubmitCallbackResponse<>(callback.getCaseDetails().getCaseData());
-            String listAssistExistsWarning = "There is already a hearing request in List assist, "
-                + "are you sure you want to send another request? If you do proceed, then please cancel the existing hearing request first";
-            response.addWarning(listAssistExistsWarning);
-            log.warn("Warning: {}", listAssistExistsWarning);
+            response.addWarning(EXISTING_HEARING_WARNING);
+            log.warn("Warning: {}", EXISTING_HEARING_WARNING);
             return response;
         }
