+++ src/main/java/uk/gov/hmcts/reform/sscs/jms/listener/HmcHearingsEventTopicListener.java
@@ -12,38 +12,36 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.jms.annotation.JmsListener;
 import org.springframework.retry.ExhaustedRetryException;
 import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.sscs.exception.CaseException;
 import uk.gov.hmcts.reform.sscs.exception.HearingUpdateException;
 import uk.gov.hmcts.reform.sscs.exception.HmcEventProcessingException;
 import uk.gov.hmcts.reform.sscs.exception.MessageProcessingException;
 import uk.gov.hmcts.reform.sscs.model.hmc.message.HmcMessage;
-import uk.gov.hmcts.reform.sscs.service.hmc.topic.ProcessHmcMessageService;
+import uk.gov.hmcts.reform.sscs.service.hmc.topic.ProcessHmcMessageServiceV2;
 
 @Slf4j
 @Component
-@ConditionalOnProperty({"feature.bypass-hearing-api-service.enabled", "flags.hmc-to-hearings-api.enabled"})
+@ConditionalOnProperty("flags.hmc-to-hearings-api.enabled")
 public class HmcHearingsEventTopicListener {
 
     private final ObjectMapper objectMapper;
 
     private final String sscsServiceCode;
 
-    private final ProcessHmcMessageService processHmcMessageService;
+    private final ProcessHmcMessageServiceV2 processHmcMessageServiceV2;
 
     @Value("${hmc.deployment-id}")
     private String hmctsDeploymentId;
 
     @Value("${flags.deployment-filter.enabled}")
     private boolean isDeploymentFilterEnabled;
-    @Value("${feature.bypass-hearing-api-service.enabled}")
-    private boolean isByPassHearingServiceEnabled;
+
 
     private static final String HMCTS_DEPLOYMENT_ID = "hmctsDeploymentId";
 
     public HmcHearingsEventTopicListener(@Value("${sscs.serviceCode}") String sscsServiceCode,
-                                         ProcessHmcMessageService processHmcMessageService) {
+                                         ProcessHmcMessageServiceV2 processHmcMessageServiceV2) {
         this.sscsServiceCode = sscsServiceCode;
-        this.processHmcMessageService = processHmcMessageService;
+        this.processHmcMessageServiceV2 = processHmcMessageServiceV2;
         this.objectMapper = new ObjectMapper();
         objectMapper.registerModule(new JavaTimeModule());
     }
@@ -55,7 +53,7 @@ public class HmcHearingsEventTopicListener {
     )
     public void onMessage(JmsBytesMessage message) throws JMSException, HmcEventProcessingException {
 
-        log.info("isByPassHearingServiceEnabled && isDeploymentFilterEnabled && deploymentId ------------------------> {}, {}, {}", isByPassHearingServiceEnabled,
+        log.info("isDeploymentFilterEnabled && deploymentId ------------------------> , {}, {}",
                 isDeploymentFilterEnabled, message.getStringProperty(HMCTS_DEPLOYMENT_ID));
 
         if (isDeploymentFilterEnabled && !isMessageReleventForDeployment(message)) {
@@ -80,10 +78,10 @@ public class HmcHearingsEventTopicListener {
                     hearingId
                 );
 
-                processHmcMessageService.processEventMessage(hmcMessage);
+                processHmcMessageServiceV2.processEventMessage(hmcMessage);
 
             }
-        } catch (JsonProcessingException | CaseException | MessageProcessingException
+        } catch (JsonProcessingException | MessageProcessingException
                  | HearingUpdateException | ExhaustedRetryException ex) {
             log.error("Unable to successfully deliver HMC message: {}", convertedMessage, ex);
             throw new HmcEventProcessingException(String.format(
