+++ src/main/java/uk/gov/hmcts/reform/sscs/jms/listener/TribunalsHearingsEventQueueListener.java
@@ -5,7 +5,6 @@ import static uk.gov.hmcts.reform.sscs.ccd.domain.EventType.LISTING_ERROR;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.jms.annotation.JmsListener;
 import org.springframework.retry.ExhaustedRetryException;
@@ -24,7 +23,7 @@ import uk.gov.hmcts.reform.sscs.service.HearingsService;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-@ConditionalOnProperty({"feature.bypass-hearing-api-service.enabled", "flags.tribunals-to-hearings-api.enabled"})
+@ConditionalOnProperty("flags.tribunals-to-hearings-api.enabled")
 public class TribunalsHearingsEventQueueListener {
 
     private final HearingsService hearingsService;
@@ -33,20 +32,11 @@ public class TribunalsHearingsEventQueueListener {
 
     private final IdamService idamService;
 
-    @Value("${feature.bypass-hearing-api-service.enabled}")
-    private boolean isByPassHearingServiceEnabled;
-
     @JmsListener(
             destination = "${azure.service-bus.tribunals-to-hearings-api.queueName}",
             containerFactory = "tribunalsHearingsEventQueueContainerFactory"
     )
     public void handleIncomingMessage(HearingRequest message) throws TribunalsEventProcessingException, GetCaseException, UpdateCaseException {
-
-        log.info("isByPassHearingServiceEnabled ------------------------> {}", isByPassHearingServiceEnabled);
-        if (!isByPassHearingServiceEnabled) {
-            return;
-        }
-
         if (isNull(message)) {
             throw new TribunalsEventProcessingException("An exception occurred as message did not match format");
         }
