+++ src/main/resources/config/application.yaml
@@ -281,10 +281,6 @@ feature:
         enabled: ${URGENT_HEARING_ENABLED:false}
     bypass-notifications-service:
         enabled: ${BYPASS_NOTIFICATIONS_SERVICE:false}
-    process-event-message-v2:
-        enabled: ${PROCESS_EVENT_MESSAGE_V2_ENABLED:false}
-    bypass-hearing-api-service:
-        enabled: ${BYPASS_HEARING_API_SERVICE:false}
     hearings-case-updateV2:
         enabled: ${HEARINGS_CASE_UPDATE_V2_ENABLED:false}
