+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/readytolist/ReadyToListAboutToSubmitHandlerTest.java
@@ -4,12 +4,17 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.openMocks;
 import static uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType.ABOUT_TO_SUBMIT;
+import static uk.gov.hmcts.reform.sscs.ccd.domain.YesNo.NO;
+import static uk.gov.hmcts.reform.sscs.ccd.domain.YesNo.YES;
+import static uk.gov.hmcts.reform.sscs.ccd.presubmit.readytolist.ReadyToListAboutToSubmitHandler.EXISTING_HEARING_WARNING;
+import static uk.gov.hmcts.reform.sscs.ccd.presubmit.readytolist.ReadyToListAboutToSubmitHandler.GAPS_CASE_WARNING;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -234,7 +239,7 @@ public class ReadyToListAboutToSubmitHandlerTest {
 
         MatcherAssert.assertThat(response.getErrors().size(), is(0));
         MatcherAssert.assertThat(response.getWarnings().size(), is(1));
-        MatcherAssert.assertThat(response.getWarnings().iterator().next(), is("This is a GAPS case, If you do want to proceed, then please change the hearing route to List Assist"));
+        MatcherAssert.assertThat(response.getWarnings().iterator().next(), is(GAPS_CASE_WARNING));
     }
 
     @Test
@@ -246,7 +251,7 @@ public class ReadyToListAboutToSubmitHandlerTest {
                 .schedulingAndListingFields(schedulingAndListingFields)
                 .region("TEST")
                 .build();
-        sscsCaseData.setIgnoreCallbackWarnings(YesNo.YES);
+        sscsCaseData.setIgnoreCallbackWarnings(YES);
         when(caseDetails.getCaseData()).thenReturn(sscsCaseData);
 
         handler = new ReadyToListAboutToSubmitHandler(false, regionalProcessingCenterService,
@@ -259,7 +264,8 @@ public class ReadyToListAboutToSubmitHandlerTest {
     }
 
     @Test
-    public void givenAListAssistCaseIfAHearingExistsInTheFutureThenReturnWarning() {
+    @Parameters({"YES", "NO"})
+    public void givenAListAssistCaseIfAHearingExistsInTheFutureThenReturnWarning(YesNo ignoreCallbackWarnings) {
         HearingDetails hearingDetails1 = HearingDetails.builder()
                 .hearingDate(LocalDate.now().minusDays(10).toString())
                 .start(LocalDateTime.now().minusDays(10))
@@ -281,14 +287,20 @@ public class ReadyToListAboutToSubmitHandlerTest {
         sscsCaseData = sscsCaseData.toBuilder()
                 .hearings(List.of(hearing1, hearing2))
                 .region("TEST")
+                .ignoreCallbackWarnings(ignoreCallbackWarnings)
                 .build();
         when(caseDetails.getCaseData()).thenReturn(sscsCaseData);
         handler = new ReadyToListAboutToSubmitHandler(false, regionalProcessingCenterService, hearingMessagingServiceFactory);
+
         PreSubmitCallbackResponse<SscsCaseData> response = handler.handle(ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
 
         MatcherAssert.assertThat(response.getErrors().size(), is(0));
-        MatcherAssert.assertThat(response.getWarnings().size(), is(1));
-        MatcherAssert.assertThat(response.getWarnings().iterator().next(), is("There is already a hearing request in List assist, are you sure you want to send another request? If you do proceed, then please cancel the existing hearing request first"));
+        if (ignoreCallbackWarnings == NO) {
+            MatcherAssert.assertThat(response.getWarnings().size(), is(1));
+            assertTrue(response.getWarnings().contains(EXISTING_HEARING_WARNING));
+        } else {
+            MatcherAssert.assertThat(response.getWarnings().size(), is(0));
+        }
     }
 
     @Test
